Rename user/views_user to user/view_user
